Move tab helper methods to app/helpers/tabs
